+++ DelvUI/Interface/PartyCooldowns/PartyCooldownsConfig.cs
@@ -17,7 +17,9 @@ namespace DelvUI.Interface.PartyCooldowns
     public enum PartyCooldownsGrowthDirection
     {
         Down = 0,
-        Up
+        Up,
+        Right,
+        Left
     }
 
     [Exportable(false)]
@@ -37,7 +39,7 @@ namespace DelvUI.Interface.PartyCooldowns
         [Order(4)]
         public bool Preview = false;
 
-        [Combo("Growth Direction", "Down", "Up", spacing = true)]
+        [Combo("Sections Growth Direction", "Down", "Up", "Right", "Left", spacing = true)]
         [Order(20)]
         public PartyCooldownsGrowthDirection GrowthDirection = PartyCooldownsGrowthDirection.Down;
 
@@ -241,7 +243,7 @@ namespace DelvUI.Interface.PartyCooldowns
                 ImGui.TableSetupColumn("Cooldown", ImGuiTableColumnFlags.WidthStretch, 10, 3);
                 ImGui.TableSetupColumn("Duration", ImGuiTableColumnFlags.WidthStretch, 10, 4);
                 ImGui.TableSetupColumn("Priority", ImGuiTableColumnFlags.WidthStretch, 22, 5);
-                ImGui.TableSetupColumn("Column", ImGuiTableColumnFlags.WidthStretch, 22, 6);
+                ImGui.TableSetupColumn("Section", ImGuiTableColumnFlags.WidthStretch, 22, 6);
 
                 ImGui.TableSetupScrollFreeze(0, 1);
                 ImGui.TableHeadersRow();