+++ DelvUI/changelog.md
@@ -1,5 +1,8 @@
 # 1.0.0.2
 Features:
+- Added "Right" and "Left" growth directions for Party Cooldowns:
+    + "Columns" renamed to "Sections". When using a vertical growth direction the Section of a cooldown would be the column. On horizontal directions, it would be the row.
+    + Due to adjustments in the positioning logic, the list might be silightly moved when updating.
 - Added an option to show mana up to 10k on Dark Knight's mana bar. Note that this will break thresholds.
 - Separated the Automaton Queen/Rook Autoturret duration tracker from the Battery Gauge into its own bar.